3397 move application-delegate screen
