+++ frontend/app/route-helpers/apply-adult-route-helpers.server.ts
@@ -126,7 +126,7 @@ export function validateApplyAdultStateForReview({ params, state }: ValidateStat
   }
 
   if (state.typeOfApplication === 'delegate') {
-    throw redirect(getPathById('$lang+/_public+/apply+/$id+/adult/application-delegate', params));
+    throw redirect(getPathById('$lang+/_public+/apply+/$id+/application-delegate', params));
   }
 
   if (state.typeOfApplication !== 'adult') {
